moved gitUser to appropriate models folder
